prepare last steps for cross country training

+++ baselines/eval_dda_adj_maps.py
@@ -108,12 +108,17 @@ def evaluate_meta_maps(map_path, template_path, wpop_raster_template):
     metadata = src.meta.copy()
     metadata.update({"dtype": "float32",
                      "compress": "lzw"})
-    with rasterio.open(hr_map_path_adj, 'w', **metadata) as dst:
-        dst.write(hr_pop_map_adj.to(torch.float32).cpu().numpy(), 1)
+    
+    if not os.path.exists( hr_map_path_adj):
+        with rasterio.open(hr_map_path_adj, 'w', **metadata) as dst:
+            dst.write(hr_pop_map_adj.to(torch.float32).cpu().numpy(), 1)
+        print("Adjusted map saved to: ", hr_map_path_adj)
 
     # reproject to the worldpop map
     hr_map_path_adj_reproj = map_path.replace(".tif", "_hr_adj_reproj.tif")
-    _, _ = reproject_maps(hr_map_path_adj, wpop_raster_template, hr_map_path_adj_reproj)
+    if not os.path.exists(hr_map_path_adj_reproj):
+        _, _ = reproject_maps(hr_map_path_adj, wpop_raster_template, hr_map_path_adj_reproj)
+        print("Reprojected map saved to: ", hr_map_path_adj_reproj)
 
 
     # define levels
@@ -151,7 +156,8 @@ if __name__=="__main__":
     Evaluates the Worldpop-maps on the test set of Rwanda
     """
     # map_path = "/scratch/metzgern/HAC/data/PopMapData/processed/rwa/buildingsDDA2_44C.tif"
-    map_path = "/scratch2/metzgern/HAC/POMELOv2_results/So2Sat/experiment_1540_88/rwa_predictions.tif"
+    map_path = "/scratch2/metzgern/HAC/data/PopMapData/raw/GoogleBuildings/rwa/Gbuildings_rwa_counts.tif"
+    # map_path = "/scratch2/metzgern/HAC/POMELOv2_results/So2Sat/experiment_1540_88/rwa_predictions.tif"
     template_path = "/scratch2/metzgern/HAC/data/PopMapData/merged/EE/rwa/S2Aautumn/rwa_S2Aautumn.tif"
     wpop_raster_template = "/scratch2/metzgern/HAC/data/PopMapData/raw/WorldPopMaps/RWA/rwa_ppp_2020_constrained.tif"
